complete chain: extract, split, synchronize
