+++ .gitignore
@@ -1,4 +1,5 @@
 **/.DS_Store
+**/*.pyc
 logs/
 data/
 .idea/
