+++ resourcesync-generator/Dockerfile
@@ -1,10 +1,7 @@
 FROM python:2
 
-
 RUN pip install resync
-COPY oai-rs/rsync.py /
-COPY oai-rs/syncdirector.py /
-COPY oai-rs/zipsynchronizer.py /
-COPY entrypoint.sh /entrypoint.sh
+COPY oai-rs/*.py /
+COPY entrypoint.sh /
 
 CMD ["/entrypoint.sh"]
