+++ resourcesync-generator/entrypoint.sh
@@ -1,21 +1,20 @@
 #!/usr/bin/env bash
 
 RUN_INTERVAL=${RUN_INTERVAL:-3600s}
-MAX_FILES_COMPRESSED=${MAX_FILES_COMPRESSED:-100}
+MAX_FILES_COMPRESSED=${MAX_FILES_COMPRESSED:-1000}
 WRITE_SEPARATE_MANIFEST=${WRITE_SEPARATE_MANIFEST:-y}
 MOVE_RESOURCES=${MOVE_RESOURCES:-n}
-
-RESOURCE_DIR="${DATA_DIR:-/input}"
-PUBLISH_DIR="${DATA_DIR:-/output}"
-SYNCHRONIZER_CLASS="${SYNCHRONIZER_CLASS:-zipsynchronizer.ZipSynchronizer}"
+SOURCE_DIR="${SOURCE_DIR:-/input}"
+SINK_DIR="${SINK_DIR:-/output}"
+BUILDER_CLASS="${BUILDER_CLASS:-zipsynchronizer.ZipSynchronizer}"
 
 while true; do
   echo "sleep $RUN_INTERVAL."
   sleep ${RUN_INTERVAL}
-  ./rsync.py --resource_dir "$RESOURCE_DIR" \
-  --publish_dir "$PUBLISH_DIR" \
+  ./rsync.py --source_dir "$SOURCE_DIR" \
+  --sink_dir "$SINK_DIR" \
   --publish_url "${HTTP_SERVER_URL}" \
-  --synchronizer_class "${SYNCHRONIZER_CLASS}" \
+  --builder_class "${BUILDER_CLASS}" \
   --max_files_compressed "${MAX_FILES_COMPRESSED}" \
   --write_separate_manifest "${WRITE_SEPARATE_MANIFEST}" \
   --move_resources "${MOVE_RESOURCES}"
