--- resourcesync-generator/oai-rs/sample/rdfdump-99999
@@ -1,1 +1,0 @@
-stopper file
