+++ resourcesync-generator/oai-rs/sample/rdfpatch-0d000000000001
@@ -1,0 +1,3 @@
+# at checkpoint   20160712144328
+# dump started    2016-07-12 14:43:28.155395
+# quad count      0
