--- resourcesync-generator/oai-rs/sample/rdfpatch-20160113072513
@@ -1,12 +1,0 @@
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#>  <http://www.w3.org/1999/02/22-rdf-syntax-ns#type>  <http://www.nanopub.org/nschema#Nanopublication>  <http://localhost:8890/jauco>
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#Association_1>  <http://www.w3.org/1999/02/22-rdf-syntax-ns#type>  <http://semanticscience.org/resource/statistical-association>  <http://localhost:8890/jauco>
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#Association_1_p_value>  <http://www.w3.org/1999/02/22-rdf-syntax-ns#type>  <http://semanticscience.org/resource/probability-value>  <http://localhost:8890/jauco>
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#>  <http://purl.org/dc/terms/created>  "2012-03-28 11:32:30.758274"^^http://www.w3.org/2001/XMLSchema#dateTime  <http://localhost:8890/jauco>
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#Association_1>  <http://www.w3.org/2000/01/rdf-schema#comment>  "This association has p-value of 0.00066, has attribute gene CENPJ (Entrenz gene id 55835) and attribute disease Seckel Syndrome (OMIM 210600)."  <http://localhost:8890/jauco>
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#>  <http://purl.org/dc/terms/rights>  <http://creativecommons.org/licenses/by/3.0/>  <http://localhost:8890/jauco>
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#>  <http://purl.org/dc/terms/rightsHolder>  <http://biosemantics.org>  <http://localhost:8890/jauco>
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#NanoPub_1_Assertion>  <http://purl.org/net/opmv/ns#wasGeneratedBy>  <http://rdf.biosemantics.org/vocabularies/text_mining/gene_disease_concept_profiles_matching_1980_2010>  <http://localhost:8890/jauco>
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#Association_1>  <http://semanticscience.org/resource/refers-to>  <http://bio2rdf.org/geneid:55835>  <http://localhost:8890/jauco>
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#Association_1>  <http://semanticscience.org/resource/refers-to>  <http://bio2rdf.org/omim:210600>  <http://localhost:8890/jauco>
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#>  <http://swan.mindinformatics.org/ontologies/1.2/pav/authoredBy>  <http://www.researcherid.com/rid/B-6035-2012>  <http://localhost:8890/jauco>
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#>  <http://swan.mindinformatics.org/ontologies/1.2/pav/authoredBy>  <http://www.researcherid.com/rid/B-5927-2012>  <http://localhost:8890/jauco>
