--- resourcesync-generator/oai-rs/sample/rdfpatch-20160113082513
@@ -1,7 +1,0 @@
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#>  <http://www.nanopub.org/nschema#hasPublicationInfo>  <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#NanoPub_1_Pubinfo>  <http://localhost:8890/jauco>
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#NanoPub_1_Assertion>  <http://purl.org/net/opmv/ns#wasDerivedFrom>  <http://rdf.biosemantics.org/vocabularies/text_mining/gene_disease_concept_profiles_1980_2010>  <http://localhost:8890/jauco>
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#Association_1>  <http://semanticscience.org/resource/has-measurement-value>  <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#Association_1_p_value>  <http://localhost:8890/jauco>
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#Association_1_p_value>  <http://semanticscience.org/resource/has-value>  "6.562110502272844e-05"^^http://www.w3.org/2001/XMLSchema#float  <http://localhost:8890/jauco>
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#>  <http://swan.mindinformatics.org/ontologies/1.2/pav/createdBy>  <http://www.researcherid.com/rid/B-5852-2012>  <http://localhost:8890/jauco>
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#>  <http://www.nanopub.org/nschema#hasAssertion>  <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#NanoPub_1_Assertion>  <http://localhost:8890/jauco>
-+             <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#>  <http://www.nanopub.org/nschema#hasProvenance>  <http://rdf.biosemantics.org/vocabularies/gene_disease_nanopub_example#NanoPub_1_Provenance>  <http://localhost:8890/jauco>
