--- resourcesync-generator/oai-rs/sample/rdfpatch-20160712144328
@@ -1,5 +1,0 @@
-# at checkpoint   20160712144328
-# dump started    2016-07-12 14:43:28.155395
-# dump completed  2016-07-12 14:43:29.325204
-# quad count      963
-# excluded graphs http://www.openlinksw.com/schemas/virtrdf# http://www.w3.org/ns/ldp# http://www.w3.org/2002/07/owl# http://localhost:8890/sparql http://localhost:8890/DAV/
